WT-11892: Handle invalid activity type code format
In some cases where the configured query may expect the code in a
different format it should not throw an exception and show the
exception message to the user e.g. the query may expect a numeric value
but since the doesActivityCodeExist() method accepts a String argument
this will result in an error and results in the exception message being
shown to the user.